+++ module/arm5e.js
@@ -438,7 +438,9 @@ function rollItemMacro(itemId, actorId) {
   if (!item)
     return ui.notifications.warn(`Your controlled Actor does not have an item named ${itemName}`);
   const dataset = prepareDatasetByTypeOfItem(item);
-  if (item.data.type == "power") {
+  if (isObjectEmpty(dataset)) {
+    item.sheet.render(true);
+  } else if (item.data.type == "power") {
     actor.sheet._onUsePower(dataset);
   } else {
     actor.sheet._onRoll(dataset);